% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob}
\alias{get_prob}
\title{`get_prob()`}
\usage{
get_prob(t = 100, lambdas, type = "series")
}
\arguments{
\item{t}{[integer] time passed. Can be a single integer or a vector of integers.}

\item{lambdas}{[vector] a vector of failure rates for components, named lambda}

\item{type}{[character] a single value describing whether these probabilities should be combined using the rules of series or parallel systems.}
}
\description{
A short description of the function!
Can be multiple lines
Notice how roxygen commenting always starts with #'
For example:
This function calculates system reliability over time t given a set of lambdas supplied.
}
\note{
- `@name` should have no spaces and match the function name.
- `@title` is the official title and can be anything.
- `@description` is a description that can span multiple lines
- `@param` marks each input parameters for your function.
- `@export` exports your function to be available to a user of your package. (Eg. not an internal function)
- `@importFrom` bundles into your package 1 or more specific functions from another package, so that your package will always function.

You can specify default inputs for an input parameter like with `type = "series"` below.

It's important to 'import' any extra functions your package relies on. 
You can use `@importFrom`. For example:
`@importFrom` packagename function1 function2
Note: if you're importing the pipeline from `dplyr`, then you can write it like I do below, where `%>%` is wrapped in ticks, eg.``.
}
\examples{

# Get series system probability at each time t
get_prob(t = c(2,4,5,6), lambdas = c(0.001, 0.02), type = "series")

# Get parallel system probability at each time t
get_prob(t = c(2,4,5,6), lambdas = c(0.001, 0.02), type = "parallel")

}
