% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_one.R
\name{plus_one}
\alias{plus_one}
\title{Plus One}
\usage{
plus_one(x)
}
\arguments{
\item{x}{(numeric) vector of 1 or more numeric values.}
}
\description{
Function to add 1 to a numeric input vector.
}
\note{
Adding `@export` below means this function will become accessible by package users, rather than being an internal-only function.
}
\author{
Tim Fraser, PhD
}
